Bug fix
Fixed bug where loading multiple experiments as sub-dirs in a folder failed due to incorrect indexing of trigger channel. Solution was to re-structure the algorithm such that DataFrames with "basic" information are created first, then averages are generated algorithmically for each experiment in that DF, and passed to the appropriate indices.